refactor * BoardDetail hooks으로 로직 이동
