+++ frontend/src/Components/Board/Board.module.css
@@ -150,6 +150,17 @@
     /*font-size: 10px;*/
 }
 
+.board_post_bottom > div > p > sup{
+    color: #ffffff;
+    background-color: #0056b3;
+    font-size: 10px;
+    padding: 2px 10px;
+    margin-left: 3px;
+    border-radius: 10px;
+    font-weight: bold;
+}
+
+
 .board_category_sub {
     display: flex;
 }
@@ -158,6 +169,17 @@
     margin-right: 20px;
     margin-bottom: 10px;
 }
+
+.board_category_sub > p > sup {
+    color: #ffffff;
+    background-color: #0056b3;
+    font-size: 10px;
+    padding: 2px 10px;
+    margin-left: 3px;
+    border-radius: 10px;
+    font-weight: bold;
+}
+
 
 .board_category_sub > button {
     padding-right: 20px;
